refactor: harden caribic startup errors
